+++ src/ui/mod.rs
@@ -1,3 +1,4 @@
 pub mod ui_legacy;
 pub mod tab_screen;
 pub mod layout;
+pub mod blocks;
